CURB-3804 Made the storefrontApiCustomerAccessToken output optional; fix Storefront API common access token handling / renewal etc.
